Create explicit class in the IR for initializer expressions. NFC
Rather than treating an init expression as a raw expressions list wrap
it into its own class.  While this does nothing now I plan to followup
by attaching an end_loc (like we do for Block) so that we can have the
validator handle checking the correct placement of the END instruction.

Its also good for readability to have an explict class for this class
of expression list.